Crate logins

Re-exports

pub use crate::migrate_sqlcipher_db::migrate_logins;

Modules

encryption

migrate_sqlcipher_db

Macros

breadcrumb

Tell the application to log a breadcrumb

report_error

Tell the application to report an error

uniffi_reexport_scaffolding

Structs

EncryptedLogin

A login stored in the database

Login

A login stored in the database

LoginDb

LoginEntry

A login entered by the user

LoginFields

LoginStore

LoginsSyncEngine

RecordFields

Login data specific to database records

SecureLoginFields

LoginEntry fields that are stored encrypted

Enums

Error

Logins error type These are “internal” errors used by the implementation. This error type is never returned to the consumer.

InvalidLogin

Error::InvalidLogin subtypes

LoginsApiError

Traits

ValidateAndFixup

Functions

ffi_logins_ab26_rustbuffer_alloc

ffi_logins_ab26_rustbuffer_free⚠

ffi_logins_ab26_rustbuffer_from_bytes⚠

ffi_logins_ab26_rustbuffer_reserve⚠

get_registered_sync_engine

Called by the sync manager to get a sync engine via the store previously registered with the sync manager.

Type Definitions

ApiResult

Result

Attribute Macros

handle_error

A procedural macro that exposes internal errors to external errors the consuming applications should handle. It requires that the internal error implements [error_support::ErrorHandling].