Struct logins::db::LoginDb

pub struct LoginDb {
    pub db: Connection,
    interrupt_counter: Arc<AtomicUsize>,
}

Fields
db: Connectioninterrupt_counter: Arc<AtomicUsize>
Implementations

impl LoginDb

pub fn with_connection(
db: Connection,
encryption_key: Option<&str>,
salt: Option<&str>
) -> Result<Self>

pub fn open(
path: impl AsRef<Path>,
encryption_key: Option<&str>
) -> Result<Self>

pub fn open_with_salt(
path: impl AsRef<Path>,
encryption_key: &str,
salt: &str
) -> Result<Self>

pub fn open_in_memory(encryption_key: Option<&str>) -> Result<Self>

pub fn open_and_get_salt(
path: impl AsRef<Path>,
encryption_key: &str
) -> Result<String>

Opens an existing database and fetches the salt. This method is used by iOS consumers as part as the migration plan to store the salt outside of the sqlite db headers.

Will return an error if the database does not exist.

pub fn open_and_migrate_to_plaintext_header(
path: impl AsRef<Path>,
encryption_key: &str,
salt: &str
) -> Result<()>

pub fn disable_mem_security(&self) -> Result<()>

pub fn rekey_database(&self, new_encryption_key: &str) -> Result<()>

Change the key on an existing encrypted database, it must first be unlocked with the current encryption key. Once the database is readable and writeable, PRAGMA rekey can be used to re-encrypt every page in the database with a new key. https://www.zetetic.net/sqlcipher/sqlcipher-api/#Changing_Key

pub fn new_interrupt_handle(&self) -> SqlInterruptHandle

pub fn begin_interrupt_scope(&self) -> SqlInterruptScope

impl LoginDb

fn mark_as_synchronized(
&self,
guids: &[&str],
ts: ServerTimestamp,
scope: &SqlInterruptScope
) -> Result<()>

fn fetch_login_data(
&self,
records: &[(Payload, ServerTimestamp)],
telem: &mut EngineIncoming,
scope: &SqlInterruptScope
) -> Result<Vec<SyncLoginData>>

fn find_dupe(&self, l: &Login) -> Result<Option<Login>>

pub fn get_all(&self) -> Result<Vec<Login>>

pub fn get_by_base_domain(&self, base_domain: &str) -> Result<Vec<Login>>

pub fn get_by_id(&self, id: &str) -> Result<Option<Login>>

pub fn touch(&self, id: &str) -> Result<()>

pub fn add(&self, login: Login) -> Result<Login>

pub fn import_multiple(&self, logins: &[Login]) -> Result<MigrationMetrics>

pub fn update(&self, login: Login) -> Result<()>

pub fn check_valid_with_no_dupes(&self, login: &Login) -> Result<()>

pub fn fixup_and_check_for_dupes(&self, login: Login) -> Result<Login>

pub fn check_for_dupes(&self, login: &Login) -> Result<()>

pub fn dupe_exists(&self, login: &Login) -> Result<bool>

pub fn potential_dupes_ignoring_username(
&self,
login: &Login
) -> Result<Vec<Login>>

pub fn exists(&self, id: &str) -> Result<bool>

pub fn delete(&self, id: &str) -> Result<bool>

Delete the record with the provided id. Returns true if the record existed already.

fn mark_mirror_overridden(&self, guid: &str) -> Result<()>

fn ensure_local_overlay_exists(&self, guid: &str) -> Result<()>

fn clone_mirror_to_overlay(&self, guid: &str) -> Result<usize>

pub fn reset(&self, assoc: &StoreSyncAssociation) -> Result<()>

pub fn wipe(&self, scope: &SqlInterruptScope) -> Result<()>

pub fn wipe_local(&self) -> Result<()>

fn reconcile(
&self,
records: Vec<SyncLoginData>,
server_now: ServerTimestamp,
telem: &mut EngineIncoming,
scope: &SqlInterruptScope
) -> Result<UpdatePlan>

fn execute_plan(
&self,
plan: UpdatePlan,
scope: &SqlInterruptScope
) -> Result<()>

pub fn fetch_outgoing(
&self,
st: ServerTimestamp,
scope: &SqlInterruptScope
) -> Result<OutgoingChangeset>

fn do_apply_incoming(
&self,
inbound: IncomingChangeset,
telem: &mut Engine,
scope: &SqlInterruptScope
) -> Result<OutgoingChangeset>

fn put_meta(&self, key: &str, value: &dyn ToSql) -> Result<()>

fn get_meta<T: FromSql>(&self, key: &str) -> Result<Option<T>>

fn delete_meta(&self, key: &str) -> Result<()>

fn set_last_sync(&self, last_sync: ServerTimestamp) -> Result<()>

fn get_last_sync(&self) -> Result<Option<ServerTimestamp>>

pub fn set_global_state(&self, state: &Option<String>) -> Result<()>

pub fn get_global_state(&self) -> Result<Option<String>>

pub fn migrate_global_state(&self) -> Result<()>

A utility we can kill by the end of 2019 ;)