logins

Module encryption

Structs

ManagedEncryptorDecryptor

The ManagedEncryptorDecryptor makes use of the NSS provided cryptographic algorithms. The ManagedEncryptorDecryptor uses a KeyManager for encryption key retrieval.

NSSKeyManager

Use the NSSKeyManager to use NSS for key management.

StaticKeyManager

Last but not least we provide a StaticKeyManager, which can be used in cases where there is a single key during runtime, for example in tests.

UniFfiTraitVtablePrimaryPasswordAuthenticator

Traits

EncryptorDecryptor

This is the generic EncryptorDecryptor trait, as handed over to the Store during initialization. Consumers can implement either this generic trait and bring in their own crypto, or leverage the ManagedEncryptorDecryptor below, which provides encryption algorithms out of the box.

KeyManager

Consumers can implement the KeyManager in combination with the ManagedEncryptorDecryptor to hand over the encryption key whenever encryption or decryption happens.

PrimaryPasswordAuthenticator

PrimaryPasswordAuthenticator is used in conjunction with NSSKeyManager to provide the primary password and the success or failure actions of the authentication process.

Functions

check_canary

create_canary

create_key

uniffi_logins_fn_init_callback_vtable_primarypasswordauthenticator