Struct logins::login::Login

pub struct Login {
    pub guid: Guid,
    pub hostname: String,
    pub form_submit_url: Option<String>,
    pub http_realm: Option<String>,
    pub username: String,
    pub password: String,
    pub username_field: String,
    pub password_field: String,
    pub time_created: i64,
    pub time_password_changed: i64,
    pub time_last_used: i64,
    pub times_used: i64,
}

Fields
guid: Guidhostname: Stringform_submit_url: Option<String>http_realm: Option<String>username: Stringpassword: Stringusername_field: Stringpassword_field: Stringtime_created: i64time_password_changed: i64time_last_used: i64times_used: i64
Methods

impl Login

pub fn guid(&self) -> &Guid

pub fn guid_str(&self) -> &str

pub fn check_valid(&self) -> Result<()>

Checks whether the Login is valid, without attempting to fix any fields. Returns an error if invalid data is found, even if it could have been fixed.

pub fn fixup(self) -> Result<Self>

Return either the existing login, a fixed-up verion, or an error. This consumes self to make it easy for callers to unconditionally replace a Login with an owned fixed-up version, preventing them from using one that is invalid.

pub fn maybe_fixup(&self) -> Result<Option<Self>>

Like fixup() above, but takes self by reference and returns an Option for the fixed-up version, allowing the caller to make more choices about what to do next.

fn validate_and_fixup_origin(origin: &str) -> Result<Option<String>>

Internal helper for validation and fixups of an "origin" stored as a string.

fn validate_and_fixup(&self, fixup: bool) -> Result<Option<Self>>

Internal helper for doing validation and fixups.

pub(crate) fn from_row(row: &Row) -> Result<Login>

impl Login

pub(crate) fn apply_delta(&mut self, delta: LoginDelta)

pub(crate) fn delta(&self, older: &Login) -> LoginDelta