Struct logins::LoginDb

pub struct LoginDb {
    pub db: Connection,
    // some fields omitted
}

Fields
db: Connection
Implementations

impl LoginDb

pub fn with_connection(
db: Connection,
encryption_key: Option<&str>,
salt: Option<&str>
) -> Result<Self>

pub fn open(
path: impl AsRef<Path>,
encryption_key: Option<&str>
) -> Result<Self>

pub fn open_with_salt(
path: impl AsRef<Path>,
encryption_key: &str,
salt: &str
) -> Result<Self>

pub fn open_in_memory(encryption_key: Option<&str>) -> Result<Self>

pub fn disable_mem_security(&self) -> Result<()>

pub fn rekey_database(&self, new_encryption_key: &str) -> Result<()>

Change the key on an existing encrypted database, it must first be unlocked with the current encryption key. Once the database is readable and writeable, PRAGMA rekey can be used to re-encrypt every page in the database with a new key. https://www.zetetic.net/sqlcipher/sqlcipher-api/#Changing_Key

pub fn new_interrupt_handle(&self) -> SqlInterruptHandle

pub fn begin_interrupt_scope(&self) -> SqlInterruptScope

impl LoginDb

pub fn get_all(&self) -> Result<Vec<Login>>

pub fn get_by_base_domain(&self, base_domain: &str) -> Result<Vec<Login>>

pub fn get_by_id(&self, id: &str) -> Result<Option<Login>>

pub fn touch(&self, id: &str) -> Result<()>

pub fn add(&self, login: Login) -> Result<Login>

pub fn import_multiple(&self, logins: &[Login]) -> Result<MigrationMetrics>

pub fn update(&self, login: Login) -> Result<()>

pub fn check_valid_with_no_dupes(&self, login: &Login) -> Result<()>

pub fn fixup_and_check_for_dupes(&self, login: Login) -> Result<Login>

pub fn check_for_dupes(&self, login: &Login) -> Result<()>

pub fn dupe_exists(&self, login: &Login) -> Result<bool>

pub fn potential_dupes_ignoring_username(
&self,
login: &Login
) -> Result<Vec<Login>>

pub fn exists(&self, id: &str) -> Result<bool>

pub fn delete(&self, id: &str) -> Result<bool>

Delete the record with the provided id. Returns true if the record existed already.

pub fn wipe_local(&self) -> Result<()>