Crate logins

Re-exports

pub use crate::encryption::NSSKeyManager;

pub use crate::encryption::PrimaryPasswordAuthenticator;

Modules

encryption

Macros

breadcrumb

Tell the application to log a breadcrumb

debug

error

info

report_error

Tell the application to report an error

trace

warn

Structs

EncryptedLogin

A login stored in the database

Login

A login handed over from the store API, which has been persisted and contains persistence information such as id and time stamps

LoginDb

LoginEntry

A login handed over to the store API; ie a login not yet persisted

LoginEntryWithMeta

A login together with meta fields, handed over to the store API; ie a login persisted elsewhere, useful for migrations

LoginFields

LoginMeta

Login data specific to database records

LoginStore

LoginsDeletionMetrics

LoginsSyncEngine

SecureLoginFields

LoginEntry fields that are stored encrypted

UniFfiTraitVtableEncryptorDecryptor

UniFfiTraitVtableKeyManager

Enums

BulkResultEntry

A bulk insert result entry, returned by add_many and add_many_with_records

Error

Logins error type These are “internal” errors used by the implementation. This error type is never returned to the consumer.

InvalidLogin

Error::InvalidLogin subtypes

LoginOrErrorMessage

LoginsApiError

Traits

ValidateAndFixup

Functions

create_login_store_with_nss_keymanager

create_login_store_with_static_key_manager

create_managed_encdec

create_static_key_manager

get_registered_sync_engine

Called by the sync manager to get a sync engine via the store previously registered with the sync manager.

uniffi_logins_fn_init_callback_vtable_encryptordecryptor

uniffi_logins_fn_init_callback_vtable_keymanager

Type Aliases

ApiResult

Result

Attribute Macros

handle_error

A procedural macro that exposes internal errors to external errors the consuming applications should handle. It requires that the internal error implements [error_support::ErrorHandling].