Crate logins

Re-exports

pub use crate::migrate_sqlcipher_db::migrate_logins;

Modules

encryption

migrate_sqlcipher_db

Macros

uniffi_reexport_scaffolding

Structs

EncryptedLogin

A login stored in the database

Error

Login

A login stored in the database

LoginDb

LoginEntry

A login entered by the user

LoginFields

LoginStore

LoginsSyncEngine

MigrationMetrics

MigrationPhaseMetrics

RecordFields

Login data specific to database records

SecureLoginFields

LoginEntry fields that are stored encrypted

Enums

ErrorKind

InvalidLogin

InvalidLoginReason

A reason a login may be invalid

LoginsStorageError

Traits

ValidateAndFixup

Functions

ffi_logins_2894_rustbuffer_alloc

ffi_logins_2894_rustbuffer_free⚠

ffi_logins_2894_rustbuffer_from_bytes⚠

ffi_logins_2894_rustbuffer_reserve⚠

get_registered_sync_engine

Called by the sync manager to get a sync engine via the store previously registered with the sync manager.

Type Definitions

Result