logins

Struct LoginDb

pub struct LoginDb {
    pub db: Connection,
    pub encdec: Arc<dyn EncryptorDecryptor>,
    /* private fields */
}

Fields
db: Connectionencdec: Arc<dyn EncryptorDecryptor>
Implementations

impl LoginDb

pub fn with_connection( db: Connection, encdec: Arc<dyn EncryptorDecryptor>, ) -> Result<Self>

pub fn open( path: impl AsRef<Path>, encdec: Arc<dyn EncryptorDecryptor>, ) -> Result<Self>

pub fn new_interrupt_handle(&self) -> Arc<SqlInterruptHandle>

pub fn begin_interrupt_scope(&self) -> Result<SqlInterruptScope>

impl LoginDb

pub fn count_all(&self) -> Result<i64>

pub fn count_by_origin(&self, origin: &str) -> Result<i64>

pub fn count_by_form_action_origin( &self, form_action_origin: &str, ) -> Result<i64>

pub fn get_all(&self) -> Result<Vec<EncryptedLogin>>

pub fn get_by_base_domain( &self, base_domain: &str, ) -> Result<Vec<EncryptedLogin>>

pub fn get_by_id(&self, id: &str) -> Result<Option<EncryptedLogin>>

pub fn find_login_to_update( &self, look: LoginEntry, encdec: &dyn EncryptorDecryptor, ) -> Result<Option<Login>>

pub fn touch(&self, id: &str) -> Result<()>

pub fn add_many( &self, entries: Vec<LoginEntry>, encdec: &dyn EncryptorDecryptor, ) -> Result<Vec<Result<EncryptedLogin>>>

Adds multiple logins within a single transaction and returns the successfully saved logins.

pub fn add_many_with_meta( &self, entries_with_meta: Vec<LoginEntryWithMeta>, encdec: &dyn EncryptorDecryptor, ) -> Result<Vec<Result<EncryptedLogin>>>

Adds multiple logins including metadata within a single transaction and returns the successfully saved logins. Normally, you will use add_many instead, and AS Logins will take care of the metadata (setting timestamps, generating an ID) itself. However, in some cases, this method is necessary, for example when migrating data from another store that already contains the metadata.

pub fn add( &self, entry: LoginEntry, encdec: &dyn EncryptorDecryptor, ) -> Result<EncryptedLogin>

pub fn add_with_meta( &self, entry_with_meta: LoginEntryWithMeta, encdec: &dyn EncryptorDecryptor, ) -> Result<EncryptedLogin>

Adds a login including metadata. Normally, you will use add instead, and AS Logins will take care of the metadata (setting timestamps, generating an ID) itself. However, in some cases, this method is necessary, for example when migrating data from another store that already contains the metadata.

pub fn update( &self, sguid: &str, entry: LoginEntry, encdec: &dyn EncryptorDecryptor, ) -> Result<EncryptedLogin>

pub fn add_or_update( &self, entry: LoginEntry, encdec: &dyn EncryptorDecryptor, ) -> Result<EncryptedLogin>

pub fn fixup_and_check_for_dupes( &self, guid: &Guid, entry: LoginEntry, encdec: &dyn EncryptorDecryptor, ) -> Result<LoginEntry>

pub fn check_for_dupes( &self, guid: &Guid, entry: &LoginEntry, encdec: &dyn EncryptorDecryptor, ) -> Result<()>

pub fn dupe_exists( &self, guid: &Guid, entry: &LoginEntry, encdec: &dyn EncryptorDecryptor, ) -> Result<bool>

pub fn find_dupe( &self, guid: &Guid, entry: &LoginEntry, encdec: &dyn EncryptorDecryptor, ) -> Result<Option<Guid>>

pub fn exists(&self, id: &str) -> Result<bool>

pub fn delete(&self, id: &str) -> Result<bool>

Delete the record with the provided id. Returns true if the record existed already.

pub fn delete_many(&self, ids: Vec<&str>) -> Result<Vec<bool>>

Delete the records with the specified IDs. Returns a list of Boolean values indicating whether the respective records already existed.

pub fn delete_undecryptable_records_for_remote_replacement( &self, encdec: &dyn EncryptorDecryptor, ) -> Result<LoginsDeletionMetrics>

pub fn delete_local_records_for_remote_replacement( &self, ids: Vec<&str>, ) -> Result<LoginsDeletionMetrics>

pub fn wipe_local(&self) -> Result<usize>

Wipe all local data, returns the number of rows deleted

pub fn shutdown(self) -> Result<()>